Refactor and centralize all auth related code
