+++ apps/store/src/pages/api/auth/login/se.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next'
-import * as Auth from '@/services/Auth/Auth'
-import { memberLoginSE } from '@/utils/auth'
+import { loginMemberSeBankId } from '@/services/authApi/login'
+import { saveAccessToken } from '@/services/authApi/persist'
 import { ORIGIN_URL, PageLink } from '@/utils/PageLink'
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
@@ -9,10 +9,10 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const ssn = req.body.ssn
   if (typeof ssn !== 'string') throw new Error('No ssn provided')
 
-  const accessToken = await memberLoginSE(ssn)
+  const accessToken = await loginMemberSeBankId(ssn)
   console.debug('Login successful')
 
-  Auth.save(accessToken, { req, res })
+  saveAccessToken(accessToken, { req, res })
 
   const nextQueryParam = req.query['next']
   const url = typeof nextQueryParam === 'string' ? nextQueryParam : PageLink.home({ locale: 'se' })
