+++ apps/store/src/services/authApi/persist.ts
@@ -1,6 +1,5 @@
 import { getCookie, setCookie, deleteCookie } from 'cookies-next'
 import { OptionsType } from 'cookies-next/lib/types'
-import { PageLink } from '@/utils/PageLink'
 
 const COOKIE_KEY = '_hvsession'
 const ACCESS_TOKEN_SESSION_FIELD = 'token'
@@ -14,13 +13,13 @@ const OPTIONS: OptionsType = {
   }),
 }
 
-export const save = (accessToken: string, { req, res }: CookieParams = {}) => {
+export const saveAccessToken = (accessToken: string, { req, res }: CookieParams = {}) => {
   setCookie(COOKIE_KEY, serialize(accessToken), { req, res, ...OPTIONS })
 }
 
 export type CookieParams = Pick<OptionsType, 'req' | 'res'>
 
-export const reset = ({ req, res }: CookieParams) => {
+export const resetAccessToken = ({ req, res }: CookieParams) => {
   deleteCookie(COOKIE_KEY, { req, res, ...OPTIONS })
 }
 
@@ -30,7 +29,7 @@ const getAccessToken = ({ req, res }: CookieParams) => {
   return deserialize(cookieValue)
 }
 
-export const getAuthHeader = (params: CookieParams = {}): Record<string, string> => {
+export const getAuthHeaders = (params: CookieParams = {}): Record<string, string> => {
   const accessToken = getAccessToken(params)
   if (accessToken) return { authorization: `Bearer ${accessToken}` }
   return {}
@@ -48,19 +47,5 @@ const deserialize = (value: string) => {
   } catch (error) {
     console.warn('Unable to deserialize session')
     return undefined
-  }
-}
-
-type LoginSEOptions = Partial<{ nextUrl: string }>
-
-export const loginSE = async (ssn: string, { nextUrl }: LoginSEOptions = {}) => {
-  const response = await fetch(PageLink.apiLoginSe(), {
-    method: 'POST',
-    body: JSON.stringify({ ssn, nextUrl }),
-    headers: { 'Content-Type': 'application/json' },
-  })
-
-  if (!response.ok) {
-    throw new Error('Unable to login')
   }
 }
